fix(emsco/user): inMyCircles by userService
This so that we can remove the appExtension injection from the dataService
