+++ EMS/core-bundle/src/Form/Data/Condition/InMyCircles.php
@@ -6,11 +6,10 @@ namespace EMS\CoreBundle\Form\Data\Condition;
 
 use EMS\CoreBundle\Entity\Revision;
 use EMS\CoreBundle\Service\UserService;
-use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 class InMyCircles implements ConditionInterface
 {
-    public function __construct(private readonly UserService $userService, private readonly AuthorizationCheckerInterface $authorizationChecker)
+    public function __construct(private readonly UserService $userService)
     {
     }
 
@@ -23,27 +22,6 @@ class InMyCircles implements ConditionInterface
             throw new \RuntimeException('Unexpected non revision object');
         }
 
-        return $this->inMyCircles($objectOrArray->getCircles());
-    }
-
-    public function inMyCircles(mixed $circles): bool
-    {
-        if (\is_array($circles) && 0 === \count($circles)) {
-            return true;
-        }
-
-        if ($this->authorizationChecker->isGranted('ROLE_USER_MANAGEMENT')) {
-            return true;
-        }
-
-        if (\is_array($circles)) {
-            $user = $this->userService->getCurrentUser(UserService::DONT_DETACH);
-
-            return \count(\array_intersect($circles, $user->getCircles())) > 0;
-        }
-
-        $user = $this->userService->getCurrentUser(UserService::DONT_DETACH);
-
-        return \in_array($circles, $user->getCircles());
+        return $this->userService->inMyCircles($objectOrArray->getCircles());
     }
 }
